Make inheritanvce relation ship between Shape.js and all other figures * All of Figure will use 'path'
